Remove typle_identity! from the most common macros
To make typle_for! work for various kinds (expr, type, stmt), a macro
was restricted to be replaced by another macro. So typle_for! was
replaced by typle_identity! which just returns its argument.

However this does not work for using typle_for! inside another macro
as the macro receives the typle_identity!() macro when it may expect a
different argument type.

Replace the macro functions with kind-specific versions to remove the
need for typle_identity!