Clippy fixes and add fn attributes to impl method
When functions are decomposed into impl's, the impl `apply` method
contains the code from the function, so needs any Clippy allow
attributes.

The attributes from the function to the impl method, so are no longer
applied to the function.

Most other attributes (rustfmt, docs) are processed before macro
invocation, so should continue to work correctly.