Support use of `typle_fold!` in expressions
When `typle_fold!` is used in an expression, it must be enclosed in
parentheses.

If there is no iteration, then the initial expression is returned, and
we need to ensure that operator precedence holds:
`typle_fold!(1 + 0; i in .. => |c| c + 1) * 3` must become `(1 + 0) * 3`.

If there is iteration, then `typle_fold!` expands to a `loop` that must
be in parentheses to use in an expression: `(loop {...; break c;}) * 3`.