Add `typle_get!` macro
The `typle_get!` macro allows a single element of the tuple to be
selected. The result is an Option, and is None if the index is out of
range. Unlike other macros, `typle_get!` can be used with a non-const
index.

Since the returned type must be uniform, a map function can be provided
to transform the selected element.